+++ src/app/components/chat.css
@@ -1,0 +1,54 @@
+.wordle {
+  grid-column-start: 2;
+  display: flex;
+  flex-direction: column;
+  box-shadow: 0px 4px 4px #00000040;
+  border-radius: 0px 10px 10px 0px;
+  width: 600px;
+  height: 808px;
+  top: 40px;
+  left: 60%;
+  position: absolute;
+}
+
+.conversation {
+  left: 30%;
+}
+
+.nameplate {
+  width: 188px;
+  height: 37px;
+  left: 37px;
+  top: 117px;
+  background-color: black;
+  color: white;
+  box-shadow: 0px 4px 4px #00000040;
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1;
+  font-size: 16px;
+  font-weight: 500;
+}
+
+.chatbox {
+  position: relative;
+  width: 364px;
+  height: 109px;
+  left: 45px;
+  top: 105px;
+  border: 1px solid #efefef;
+  filter: drop-shadow(0px 4px 4px #00000040);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 14px 26px;
+}
+
+.portrait {
+  z-index: 2;
+  left: 410px;
+  top: -20px;
+  position: relative;
+}
