+++ src/app/components/picture.css
@@ -1,12 +1,20 @@
+.main {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  position: absolute;
+  width: 1200px;
+  height: 808px;
+}
+
 .banner {
   display: block;
-  width: 500px;
+  width: 724px;
   height: 100px;
-  left: 45%;
+  left: 71%;
   text-align: center;
   background: transparent;
   position: absolute;
-  top: 20px;
+  top: 5px;
   z-index: 1;
   border-left-color: transparent white;
   clip-path: polygon(45px 0, 100% 0, 100% 100%, 0 100%);
@@ -17,45 +25,55 @@
   content: "";
   position: absolute;
 }
+
 .banner:before {
   display: block;
   width: auto;
-  height: 50px;
+  height: 83px;
   background: #000000;
   position: relative;
+  box-shadow: 0px 4px 4px #00000040;
 }
+
 .banner:after {
   position: relative;
   height: 0;
   width: 5px;
-  right: -245px;
-  bottom: 10px;
+  right: -358px;
+  bottom: 38px;
   border-top: 10px solid #000000;
   border-right: 10px solid transparent;
 }
+
 .tittle {
   color: white;
   position: relative;
-  top: -35px;
+  font-weight: 600;
+  font-size: 32px;
+  line-height: 48px;
+  display: flex;
+  align-items: center;
+  text-align: center;
+  top: -66px;
+  left: 30%;
 }
-.background {
+
+.poke-view {
+  grid-column-start: 1;
   display: flex;
   flex-direction: row;
-  max-width: 900px;
-  position: absolute;
+  position: relative;
   top: 5%;
-  left: 13%;
+  left: 205px;
+  width: 600px;
+  height: 808px;
 }
-@media only screen and (max-width: 480px) {
-  .background {
-    display: flex;
-    flex-direction: column;
 
-    position: absolute;
-    top: 5%;
-    left: 10%;
-  }
+.background img {
+  border-radius: 10px 0 0 10px;
+  box-shadow: 0px 4px 4px #00000040;
 }
+
 .run button {
   border: none;
   padding: 5px;
@@ -63,13 +81,14 @@
   border-radius: 20px;
   padding: 8px;
   font-size: 20px;
-  box-shadow: 0 3px rgba(0, 0, 0, 0.233);
+  box-shadow: 0 3px #0000003b;
   position: relative;
-  top: 70px;
-  left: 220px;
-  width: 180px;
+  top: 40px;
+  left: 200px;
+  width: 155px;
 }
-.run img {
+
+.run svg {
   -moz-transform: scaleX(-1);
   -o-transform: scaleX(-1);
   -webkit-transform: scaleX(-1);
@@ -77,6 +96,7 @@
   filter: FlipH;
   -ms-filter: "FlipH";
 }
+
 .pokemon-images {
   display: block;
   position: absolute;
@@ -84,12 +104,27 @@
   max-height: 300px;
   justify-content: center;
   align-items: center;
-  filter: drop-shadow(0 0 rgba(0, 0, 0, 0.404));
+  filter: drop-shadow(0 0 #00000067);
   top: 50%;
-  left: 35%;
+  left: 55%;
 }
+
 .hidden {
-  filter: contrast(0) brightness(0);
+  filter: contrast(0) brightness(0) blur(25px);
+  max-width: 350px;
+  max-height: 350px;
+}
+
+.hidden1 {
+  filter: contrast(0) brightness(0) blur(20px);
+}
+
+.hidden2 {
+  filter: contrast(0) brightness(0) blur(10px);
+}
+
+.hidden3 {
+  filter: contrast(0) brightness(0) blur(5px);
 }
 
 .oval-shadow {
@@ -101,17 +136,4 @@
   max-width: 300px;
   bottom: 90px;
   z-index: -1;
-}
-@media only screen and (max-width: 480px) {
-  .pokemon-images {
-    display: block;
-    position: absolute;
-    max-width: 300px;
-    max-height: 300px;
-    justify-content: center;
-    align-items: center;
-    filter: drop-shadow(0 0 rgba(0, 0, 0, 0.404));
-    top: 50%;
-    left: 25%;
-  }
 }
